Separate tests for individual array and object modules. Remove test for prime/mixin.js. Update prime tests. Update Makefile test path.
